[FIX] sale: pricelist price base
Fine tuning of 1a7f2d6909354b964af45cd892b84bfdbb803daa

When setting a pricelist to show the public price discount, it ignored
the pricelist item's base field (if not 'Other Pricelist') and set it
by default on the lst_price instead.

This did not work when you based your discount on the standard_price
(= cost price) with public price discount.

Now it does.

opw:775275