[IMP] anlaytic: display total of balance/credit/debit on group_by
Purpose
=======

User not able to see his total balance per customer if user has several projects with same customer..

see here : https://www.odoo.com/web/content/3287924

Specification
=============

User should be able to see the sum of the debit, credit and balance when group by partner

The difficulty here is that debit,credit and balance are non-stored computed fields depending on the company and the date passed with the context. As these parameters depend on the user and the current date, we can't store the field. The simplest way to achieve this is to override the read_group method and compute the sum by hand when the values are aggregated (means when a group_by is applied).