[REV] base: revert "[IMP] base: no image (and kanban use default values"
This reverts commit 70fbfa9b67c3fb8647cd41b4710220141c60d717.

It is not clear why it has been done. Original author thinks it can
be reverted as it makes Odoo have people without image which is not
good when starting with Odoo. We don't know what we want but what is
certain is that we need images instead of ugly void placeholder.