[IMP] purchase: edit received quantity on purchase_line when type service
When we purchase products of type services,the received quantity is
always set to the ordered quantity.

We want to be able to manually enter a received quantity on service
products to have a better trace of what have been received or not.

We show the picking stat button olnly if there are picking

We show the columns qty_received, qty_invoiced and invoice_status
depending on the parent state, and not depending on the context passed
in the menu item.