[REF] purchase_requisition: blanket order generate sellerinfo
The blanket order is a contract between you and a supplier on a
different
price (usually lower than normal) for a big quantity of some specific
products.

The goal is to generate a supplier info for each product of the blanket
order and
remove them when the BO is either closed or cancel.