[FIX] point_of_sale: take into account extra variant prices
list_price is the base price of a product which does not include
price_extra.

Note that in the backend the pricelist computation uses

product.price_compute('list_price')

to compute the base price on which to apply the pricelist. For
product.product this seems to be equivalent to the lst_price computed
field.

The original list_price field is unused but has not been removed to
avoid breaking any 3rd party code that is using this field.

opw-781629