[FIX] point_of_sale: some events changed to jquery
Before this commit, it was not possible to type on a physical keyboard
the amount of a payment in the relevant pos screen

This was due to the refactoring of BarcodeEvents (commit 464110a019a8923fc0ec991a697814ef3dd63e0a)
which converted vanilla JS events to Jquery ones, which the pos was not ready for.

After this commit, it is possible to type the amount of a payment
in the payment screen, just as it worked in 10.0 so far.

OPW 781242

closes #21094