[FIX] project, sale_timesheet: analytic account and project creation
When we create an analytic account, a project is automatically
generate, and it should not. This is a bug coming from
the removal of 'use_tasks' field. See 753561e218b918db8fa65ed2cfb86011348f4af5.

We can now remove all the 'project_create' and
'project_creation_in_progress' context key.