[IMP] account: allow change 'reconcile' field on account.account
That operation was previously forbidden because it could be potentially costly as we may end up having to recompute the 'amount_residual' and 'amount_residual_currency' fields on a lot of account.move.line.

Basically, we now authorize it and manually handle the update of that field via a manual SQL query.

Was PR #19204. Was task 32125.