[FIX] barcodes: manage several barcode widgets
related to rev b0878ff96eab220faa243ab70cf0ca44fd6dc1d6

It was broken following the refactoring done in saas16, resulting in
barcodes being handled on all the displayed forms and it was problematic
in the case of a form dialog openned over a "fullscreen" form view.

This was particularily problematic for the scrap dialog in the
stock_barcode module.