[FIX] mrp: workorder SN required only when moves are consumed
When the user do a work order with a final product
tracked by lot or serial number. A serial number is
asked at the end of the first workorder. The purpose is
to link consumed products with the SN/lot produced.

However it could happens that no component are consumed
during the first and following work order.

This commit allows to by pass the lot/SN registration if
nothing is consumed.