[FIX] web: restore saas modals
This commit basically reverts 3105ecfaa10ae228dec57fd4f7030135c03780af
The fix was unstable (as restoring a feature that was not there
before release, even if originally merged before). This was in
particular breaking saas modals as lots of them are still opened
using 8.0 system... and opening them this way might make them appear
below their own backdrop in the new system the mentioned commit
restored.

We will think about that feature again for a future version.