[FIX] web_editor: revert copy/paste feature of the editor
The feature was introduced and fixed by following commits:
- https://github.com/odoo/odoo/commit/58477125114154827e0041bf288a5eaec3cfe658
- https://github.com/odoo/odoo/commit/7de5ee7f0cd199f7794ad7bea8165d90c6d4b0b8
- https://github.com/odoo/odoo/commit/f213b6284d79ae35d4cd16b8ff049b5dfb1c4513

The feature is however still buggy in lots of cases as it is a bad
heuristic that cannot handle simple pasting cases. The feature was
initially introduced to allow pasting PDF content without having to
remove all the unwanted line breaks this induces. As this feature is
not that important and introduced more bugs than it solved, the feature
is deleted by this commit. If we want that feature again, maybe it is
worth considering an "Import from PDF" button or something like that.

See https://github.com/odoo/odoo/issues/15665