invoice line tax fix
The Odoo 'Entries analysis' (cf. Reporting -> Accounting -> Entries Analysis) gives wrong
results for the 'Product Quantity' measure for intracom acquisitions with tax config based on tax codes
(e.g. l10n_be, l10n_es)

In this case the intracom acquisition expense needs to be reported on 2 'base' tax codes.
the account.invoice.line,move_line_get method creates in such instances an extra
account_move_line and nullifies the 'price' and 'account_analytic_id' values but not
the quantity field.

The result is that the 'Entries Analysis' report shows twice the actual number of acquired products.

This is fixed by also putting the 'quantity' to zero in this method.