[FIX][IMP] sale: set sale_warn as non required
When reinitializing modules it is impossible to create partners if
sale is installed because sale_warn field has a required column
but no default value if the module does not depend from sale. Use
case is trying to run test_mail tests with sale installed.

As this field is not business critical it is now not required. As there
is a default value behavior should not change for users.

See also 17f6a59561179ea6207394aa58efec910bde09a7 and 98e09246cf7d4a8973c5314a67d75c3314490657.