[FIX] website_(event_)sale: fix tours in enterprise
Recently, a few tours started failing in enterprise.  The reason is that
there is some code in the e-commerce which folds the navigation bar
(with the page names), including the shopping cart counter.  Some tours
depended on the presence of that counter, and fails in enterprise, when
more modules are installed and the folding code is more likely to
actually run.

In this commit, we do not fix the core issue (not hiding the cart
counter), but we work around the issue by slightly changing the tours to
make them pass.  I believe that the website team will make that properly
work soon.