[FIX] doc: informal sub-params
Revert 6bdd7edd1308dfe070c82c4924c6af33c563340d which reverted
81666425d24796a672acaad7f21fbea71d912e41.

Issue was the support for sub-parameters (aka @param {type} foo.bar)
where the "parent" parameter was not formal (only present in a
function's docstring, but not in the function's parameters list),
subtype creation would fail to find a root parameter to attach the
sub-parameter to.

This has not actually been fixed, but for the specific case where the
formal parameters list is *completely empty* we fall back on a list of
the docstring's parameters, which works well enough for this case.