[FIX] point_of_sale: automatic reconciliation with returns
In the case where:
1- Do an order on the pos selling a product A (amount = 10)
2- Do another order returning a certain quantity of A (with a quantity < 0)
3- Do a third order selling something else (amount = 5)

Before this commit, the account move lines of the payment move and of the sale move of operation 2 weren't reconciled
This was because the credit to the receivable account of the return was reconciled with the debit of the first operation,
leaving the debit of the payment of the return alone
(i. e. the lines number 1 & 3 & 5 were reconciled together, instead of 1 & 2 & 5, and 3 & 4)

We end up in a situation like this (at validation time)

	RECEIVABLE
D	|	C	||	OPERATION		|	Line Number
----------------------------------------------------------------------------
15	|		||	(1 & 3-Engaging Sale)	|	1
	|	10	||	(1-Payment)		|	2
	|	10	||	(2-Engaging Return)	|	3
10	|		||	(2-Payment Return)	|	4
	|	5	||	(3-Payment)		|	5

After this commit, we reconcile the returns for each order before reconciling other transaction chains
The fix is limited in its scope as the automatic reconciliation for the pos transaction is fairly recent
and a "nice to have"

OPW 787298