[FIX] web: ActionManager: honor action's groupBy
Before this rev., the group_by specified in the action context was
completely overridden by the group_by specified in the search view,
even if it was []. In other words, it was completely ignored.

Moreover, it was only took into account (before being overridden)
at the first load of a view, not at reload.