[FIX] google_drive: tests: hasSidebar param
Rev. 0104d2a0 fixes the google_drive override of the Sidebar in
saas-16. The test specifies param 'sidebar: true', so that a
sidebar is instantiated by the FormView. This param has been
renamed to 'hasSidebar' by rev. 61c7645 in master, so the test
broke when rev. 0104d2a0 has been forwardported to master, and it
thus needed to be adapted.