[FIX] sms: switch `phonenumbers` warning to INFO
Having it in INFO should be sufficient for its purpose, and will avoid
impacting all CI builds done on a system that does not have the lib
installed.

For the record, this is not a hard requirement because the lib was not
available in Debian stable packages at the time of release. It is only
enabled on demand for those who want the feature and can install it
manually.

Also slightly improved the message.

Fixes #22426
Closes #22459