[FIX] mail: livechat: send message with chat_window
Revision on 48ff0da58166db2ba828351d6e53571fd48b1cc9
The changes on chat_window are wrong, because the
composer uses an 'input', instead of a 'div'.

As a result, it was not possible to send messages
over im_livechat from the website.

This was not the case on the webclient: chat_window
is enhanced with extended_chat_window, which replaces
the composer with 'div', hence the changes are correct.