[FIX] website_forum: allow portal users to post replies to replies
Before this commit, when connected as a portal user, the user was not
able to post replies to existing replies. This is because the textarea
section was marked to be displayed for connected internal users only.
This commit removes the whole group restriction. Indeed, it makes sense
(as a fix), to show the textarea section for public users too. Indeed,
the "Reply" button is already shown to public users but does not do
anything. With this commit, clicking the "Reply" button shows the
textarea section for everybody and, if the user is not connected,
redirects to the login page when posting the message. This of course
needs usability improvements in master.

Note: the problem had originally been fixed for regular answers with
commit https://github.com/odoo/odoo/commit/6931e7b06fe2bb714405a2f32d9ec4fa25825b49

Note 2: both mentioned problems were originally introduced by commit
https://github.com/odoo/odoo/commit/3d197ab9a2e655fd5095d4a605bb11395bc7c5b4

Closes https://github.com/odoo/odoo/issues/22648