[IMP] web: display no content helper in grouped kanban view
Before this commit, it was not possible to display a help message in a
kanban view, if it is grouped.  This is an issue, from an onboarding
point of view.  Imagine a new user, creating a project.  After it
creates a few columns, it is not clear what to do.

With this commit, we display the nocontent help message in a grouped
kanban view, if the user is not currently creating a new column. This
allows the user to focus on its work, by working in phases: first, he is
prompted to create project column, then when he decides that he is done,
the user interface shows the nocontent help message, which guides
her/him to the process of creating a task.