[FIX] web_editor: allow deleting all images in table cells
Before this commit, it was impossible to delete an image which was a
direct child of a table cell using the trash-icon button in the
toolbar.

This commit fixes the behavior by changing the logic quite a lot, so
it is kinda tricky (see https://github.com/odoo/odoo/commit/8a017e041915e5f9792c044809b34bcbf41e6c59
for more explaination).

Closes https://github.com/odoo/odoo/issues/22523