[FIX] web: hide nocontent helper in grouped kanban with no column
Before this commit, the no content helper was displayed in grouped
kanban without any column.

It should be hidden, because the user expects to have a button to
create a task, which is intentionally hidden in this case.
(To sum up, the user is expected to create new columns before being
prompted to create new tasks)