[FIX] website_sale: data > demo data
Do not rename a translatable record in demo data.
Renaming a translatable record will breaks translations.
I am aware this renaming was done for demo purposes but it is not a good enough
reason to force people in another language (e.g. Chinese) to have the main
pricelist to be named 'USD'.

Exporting the translations only when the product module is installed is not an
option as many product descriptions are added in website_sale
(website_description field) and would then be missing from the translated
content.

It is the same problems as renaming menu based on module installation, it
will not work for non-English speakers.

Will also avoid manual corrections of pot file such as in #23686