[REM] hr_expense: remove assigned subtype and keep only classic tracking
Since 7761c4353ff44b5308f8346e4bb630133f58602d assignation is now a notification sent in the
inbox or by email to the newly-assigned responsible. It is therefore
considered as not necessary to link a subtype to the assignation itself.
Indeed having it logged in the discussion history as a classic tracking
in addition to the notification is considered as sufficient.

This commit is related to task ID 60790. Closes #23244 .