[FIX] pos_mercury: don't attempt negative transactions
Normally Vantiv should return an 'Invalid Purchase Amount' error. But
sometimes they will process the transaction anyway as if it had a
positive amount. We're sure the amount we send is negative. After
communicating with Vantiv it's unclear why that sometimes
happens. They recommend to explicitly prevent negative charges and to
not rely on them throwing an error.

Since we have to display an error message to the user anyway we take
the opportunity to explain the correct full flow to handle refunds.

opw-805302