[FIX] account: false complete_tax_set tax default
When setting a chart of account on a company, we have default taxes from
the default chart of account different from chosen one.

But if a chart of account is chosen, it should override them in the
onchange_chart_template_id function. This works as expected for a
truthy complete_tax_set chart of account template, but not for a falsy
complete_tax_set chart of account template.

The complete_tax_set indicates that we don't want to set default taxes
and the user should do it or modify the given tax rates himself.

opw-1824751
closes #23655