[FIX] website_sale: dont show variant option when not active
Before this commit, a user that ccick 'Customize' on th eproduct page
see the 'List View of Variants' aven if option 'Variants' in settings
was disabled.

Now we hide this view if the group for variants is not enabled in the settings.

This commit closes #23398

That means that a user that enable the view while the variants was enable,
need to disable the view before disable the variants.