[IMP] delivery: Improve the SO invoice status
Purpose
=======

If a SO is only containing lines for which the product is invoiced on
the delivered quantity, the sales order is marked as 'Nothing to Invoice'.

If I add a delivery line, the SO becomes 'To Invoice'. There is not really
a use case in the real life where you will invoice the delivery fees before
invoicing what has been really delivered.

Specification
=============

If an order:
- contains delivery lines or downpayment lines
- If the product on the other lines are invoiced on the delivered quantity
- Nothing has been delivery or can be invoiced for theses lines

Set the invoice status of the sales order to 'Nothing to invoice' instead
of 'To Invoice'

Task ID : 33126