[FIX] account: make Journal Item company read-only
PR #16826 (rev 9741a3a9d6534a7804a84ec5162c6959ee8f2664) added the
`company_id` field in the Journal Entry form.

Unfortunately that field was not marked read-only, so creating a move
would write to this field, triggering a useless and possibly very long
recomputation of many related fields based on the company of the
account. It must be read-only, its value solely driven by the value of
the account_id field.