[FIX] account: fix wrong domain/default journal on account payment view.
Fixed use case:
When creating a new payment (not from invoice), the payment amount is zero and this value triggers an onchange method that set the first miscellanous journal_id as the payment journal_id and force the domain with only miscellanous journals.

This behavior was added in commit 5b310c7 to be able to clear invoices by registering a 0 amount payment on it, but this has obviously only interest when registered from the invoices form view... To keep this functionality we only add a "AND" sentence to validate that the payment has invoices, if the payment has not any invoices the journal and the domain of this field will be for bank and cash.

Was task ID 1823023. Was PR #23609. Nice courtesy of Oscar Ulises Garza Córdova