[FIX] hr_timesheet: analytic account generation
'allow_timesheet' is true by default, but when installing hr_timesheet for already
created project, populating the column does not use write method, and so, no
analytic account is generated for existing project.
It is required to use the same mecanism as in resource (calendar generation),
and to do so, rewrite part of create/write method of project.project.

This commit also add an onchange, to maintaint consistency about having
no analytic account does not allow timesheet on project.