[FIX] website_sale_wishlist: correct tour for new checkout
The website sale wishlist tour checks if you are correctly redirected to the
cart.
To do so, it checks if the cart title 'Shopping Cart' appears on the screen.

But since there is a new checkout in saas-11.3 with ae620553c that removed this
title, we need to trigger on something else.