[FIX] website_sale: make automatic invoicing more usable
In V11 with automatic invoicing set in website settings, it's no longer
possible to confirm an online order paid with wire transfer without
registering the payment. A user might want to confirm the order and then
register the payment.

Deactivates the `Mark As Paid` button on the order. (by removing the
xpath)

Task #1819491
Closes #24347