[FIX] purchase: use PO company for PO notification
In cf43b572 the "Invoice Notification Email"  and "Sale Order
Notification Email" were modified so the company used was the one of
the object if available or the sender otherwise.

This had not been done for the "Purchase Order Notification Email" which
was added in saas-14 which might be unexpected.

Also as in 58ccac833 specify the company when displaying the logo.

opw-1838112
closes #24344