[FIX] web: ActionManager: store evaluated context
... in session_storage.

The current action is stored in the session_storage (stringified),
so that it can always be restored on F5 (useful for dynamic
actions). Before this rev., it was stored with it's 'raw' context,
i.e. before its evaluation. However, when this context is an
instance of a CompoundContext, it is meaningless once restored.

So this rev. ensures that the action is stored with its evaluated
context.