[FIX] web: fetch name_gets on list only when there are data
Revision on 7b3877c36f61de6e1b399f62360569955bad9bdc

The commit above does not check whether the list contains data or not.
As a result, if there are no data in the list, there is no reason to
fetch `name_gets`.

In fact, there could be a traceback because some parameters of this rpc
are computed from data in the list. So if there are no data in the list,
the parameters of the rpc are undefined. This was the case for a test
in enterprise.