[FIX] base: rng tree def: js_class must be included
Before this commit, the js_class attribute on a tree arch was not present in the RNG of the tree view.
This is inconsistent with what is done in Odoo, where that attribute is used for kanban views,
is defined in the view_manager to select the widget,
yet unusable for other views.

After this commit, we add this attribute in the RNG to be able to use it in tree views

Linked to OPW 1837565
closes #25088