[FIX] sale_timesheet: fix view mode
Actions returned by the plan project are directly
executed by the webclient. But their are not post
processed like other actions by /web/action, so the
'fix_view_mode' method is not call, and this causes
a mismatch between 'list' and 'tree' mode and type
on action.
This 'hack' is a bad legacy of the past ...