[FIX] purchase_stock: correctly set dropshipping settings
Currently settings for purchase stock dropshipping are added by
classname 'app_settings_block'. But there are so many classes with this name
that by default it takes first class. First found application receives this
settings.

This commit fixes that by correctly putting the settings in the purchase
settings.

Related to task: 1858217
Closes PR #25267.