[IMP] account: update dashboard sales and purchase journal graph
The sales journal graph  should show the cash-in forecast of invoices.
In order to do that we need to display residual amount to be received
based on the invoice due date.

The purchase journal graph should show the cash-out forecast from
invoices. In order to do that we need to display the residual amount to
be paid based on the invoice du date.

Was task: 1855081
Was PR #25344