[FIX] mrp: bom cost multiplied twice
Use case to reproduce:
- Create a BoM with a bom line that have at least 2 quantity to process and
this line should be a manufacturable product with also a BoM.
- Open the BoM cost report.
The line has its correct BoM cost multiplied by the quantity to consume.
It happens because the price of the BoM line is computed with the right
quantity to consume however it's result is multiply one more time to the
quantity to consume.