[FIX] stock: be able to update stock module (demo data)
Due to commit e3207a9582e79c4eb01e587ab8be53f3f26942c5 when a user
with demo data try to update stock it will return the location
constraint because the demo data try to update the supplier location
with the usage supplier. However there is no check if the usage is
really modify or just in values.

This commit add a check in order to only select locations with a new
usage or that effectively change the 'scrap location' switch.