[FIX] base: properly render settings form view after reload
Before this commit, something strange could happen after reloading a
settings form view: all 'sections' were displayed.  The issue was caused
by the fact that the state of the view was read from the arch of the
form view at render time.  Also, as a form of optimization (I suppose), the
settings did not call the initModules method twice, when it rerenders.
This is unfortunate, because the very same method is actually
responsible for removing unused sections.

This fix is the safe fix, but definitely not the perfect fix (this would
involve a larger change, something like reading the state from the arch once in
the view (factory), then only rendering what is necessary).