[FIX] mail:tests: more precise jQuery selector on modal body content
Tests from [1] were not passing, due to changes from [2].

This is due to the fact that `$.('.modal-body')` picks 2 elements, instead of 1.

Partial HTML before [2]:

```
    <main class="modal-body">
        <div>
            <p>
                ...
            </p>
        </div>
    </main>
```

Partial HTML after [1]:

```
    <main class="modal-body">
        <main role="alert" class="modal-body">
            <p>
                ...
            </p>
        </main>
    </main>
```

In order to not pick both DOM elements, we refine the jQuery selector to pick
the `<p>`, so that there's only one element and it has the same text content.

[1] [FIX] mail: warn user on record reload before post message
    https://github.com/odoo/odoo/commit/cf312f1edd326c9fe01c7157f6453678f33c9084
[2] [IMP] Improve compatibility with screen readers (accessibility) (#24574)
    https://github.com/odoo/odoo/commit/9784d7e22a9bc83f6cb9daf8ee0e72e6d80d6bbe