[IMP] mail: clean record name displayed in notification email and update font size
Purpose is to avoid having '/' in record name. This is generally considered
as being an URL and leads to a specific display in most email readers. In
this commit we therefore remove '/' by '-'.

A use case is the name of payment receipts.

Font sizes in light notification template are also standardized to 13 and
11 px to better match the chatter layout. Emails are still bioutifoul.

This commit is linked to task ID 1843376 (and 1868112) and to PR #25349
(and #25889).