[FIX] web: make the kanban view wait on async field widgets
Before this commit, the kanban view ignored the fact that a field widget
could be asynchronous.  Worse, some preliminary support was done in
kanban record, but this would crash if an async widget was used.

With this commit, we properly synchronize the kanban
records/columns/renderer.  If for some reason a widget needs to be
added to the DOM before it is ready, it can simply returns an empty
value in the willStart/start methods.