[FIX] mail: expand/collapse planned activities in chatter
Before this commit, when the user clicks on the chatter category
"Planned activities", the user is redirected to the home menu.

The "Planned activities" category in the chatter is only visible when
there is at least one planned activity on the record.

The intended behaviour is to expand/collapse the planned activites on
the chatter. This commit fixes this issue.

Task-ID 1864452