[IMP] sale_timesheet: hide 'invoice' button when no service to invoice
In plan project, when there is something to invoice on the SO
the "Create Invoice" button will be displayed, even if the
"to invoice" amount is zero. This can happen when the SO contains
an ordered service line (that does not create project/task). This
line will be in 'to invoice' status, but it is not taken into
account on the plan project since it does not interact with the
project (no created task, no timesheet, ...)
So, to avoid confusion, we hide the button on the project
overview.

Task #1857652