[IMP] auth_signup: Set activation_scope at 'B2C' by default
Purpose
=======

More and more features redirect to the portal (see & pay
quotations, invoices, etc.) with anonymous access by default
you can't sign up to the customer portal, while you see
a Sign in button. It's frustrating.

The ability to create a customer account is a nice service,
useful in any business willing to engage with returning
customers. This should be the default value

Specification
=============

Activate B2C portal by default. We have to create an
ir.config.parameter for the change to be immediate after
the installation. Otherwise the admin has to go and save
its res.config.settings one time before it's effective.

Task ID: 1841514