[FIX] point_of_sale: support non-company currencies
Multiple currencies in the same POS have never been
supported. However, before 5e54b1b568fbae826e11211becb0fea44c5d0605 it
was possible to load a POS with a currency different from the
company's currency through the sales journal.

This re-introduces that option by manually converting to the default
pricelist currency if it's different from the company currency.

The currency conversion implemented here is based on (and should be
identical to) _convert on res.currency.

opw-1863047