[FIX] hr_holidays: remove double label and fix default value
Default value of interval_number and number_per_interval was overriden
and set to 0 by _onchange_holiday_status_id this would then crash at
creation because of the sql constraint.

In the form view of allocations, the duration label was duplicated,
solution was to remove labels and simplify the view.