[FIX] account: partially revert 0f73cfbb
The changes should not have been merged in stable.
Users installing the account module before 0f73cfbb had a not null constraint
on the field company_id.

After updating the code, the field is still required but has no default value.

If a chart of account is installed after updating the code, it would fail.

opw-1867358