[FIX] mail: post message with emojis x'D (😂)
Before this commit, when posting a message with some emoji sources (e.g. "x'D"),
sometimes it was not transformed into the corresponding unicode representation
(e.g. "😂").

The emojis with this issue have different escaped and unescaped source
representations:

  - (unescaped) "x'D"
  - (escaped) "x&#x27;D".

The emoji converter was mistakenly using the escaped representation, instead of
the unescaped one.

Task-ID 1868892