[FIX] web: look into the DOM for loadCSS
Since rev. odoo/odoo@618e4b3 loadCSS function in ajax returns a promise
in order to known when the stylesheet is loaded.

It uses a local cache to avoid fetching the file multiple times.
However, the lazy loaded lib could already be in the DOM (in the
test bundle for example). In this case, the file was fetched twice.

In this case, the cache is now filled with a resolved promise.