[FIX] models.py: onchange creating m2m => no NewId
Since fed2953 if we created a record in a many2many relationship of an
onchange, we would also possibly erreneously return the NewId.

The logic of convert_to_onchange of _RelationalMulti doesn't send back
the ID field (if the ID is needed, it is part of the command, not the
update/create dictionary values of a command) and this commit does the
same for the code changed in fed2953.

This commit also adds a test for this fix, and a test for the regression
corrected with 678737fd5d reported in #25969

fix for https://github.com/odoo/odoo/commit/fed2953b0#commitcomment-29837226
closes #25984