[REF] point of sale: revert wrong configuration change
This commit reverts a configuration update recently done. Tax display
configuration was moved from pos config to general company settings.
This was a mistake and should not stay as it is. Even if saas 11.4 could
be considered as stable this change is necessary and does not impact
anybody as this branch is not released.

Partially revert 1a03a447a9a52d1b6ada37d047ed209e8db7a010 .