[FIX] website_sale: _get_pl_partner_order
With a9e63d4, people can use promo code on website.
But the condition leads to a side effect, if a partner has a pricelist set for
him, he will never be able to see the dropdown list of pricelist since he will
enter the else condition with show_visible set to true.

Now, user is able to see selectable pricelists.

opw-1869371