[IMP] mail: create tracking value using the insert multi
Purpose of this commit is to speedup tracking values creation by manually
creating them using the newly-introduced multi create. As tracking values
creation is already done manually as sudo when creating mail message we
just update the code to directly insert them as multi.

This commit is linked to task ID 1870656 and PR #26068. No functional change
should occur with this commit.