Revert "[FIX] models: make `onchange` return a smaller diff"
This reverts commit fed2953b080a981cf68bc7d715b7d174bc68c235 which
already has been fixed multiple times [0].

However this change does not correctly handle successive onchange calls on
ony2many fields which alter existing rows.

i.e., in saas~11.4, an onchange has been added on move lines [1] in
order to create and/or update lines containing taxes amounts.
However, with diff-only onchange, a second call to onchange will discard
changes made by a first one as nothing change between the 2 calls.

Scenario: On an existing account.move, add a new credit line with a tax set.
The first onchange will update the tax line with new amount. Add a
second line (with debit amount balanced). The onchange will not change
the tax line, resulting as it value being reset to initial value. As a
result, the move is not balanced and cannot be saved.

[0] 678737fd5dc65c57432251bdc39ccfdc6317444f, dd69c7636883817f1b495dd203954730f5795ab2, eaad8fbb64ee230e48db2d775c778ae8d585b7fd
[1] c447b446d3a71df611de0c80cbf7ece58af25e4c