[FIX] barcodes: ok PageUp/Down/Home/End in firefox
Same reason as 83bd34c5.

In firefox if barcode is installed, for example on a form view the keys
to navigate the possible scrolling PageUp/PageDown/Home/End don't work
at all.

This is caused by the interface barcode catching and differences of the
key event handling between firefox and other browser.

So in this change, we add these keys so they are not catched by barcode.

note: aso backport 87e8aa177 so the code is the same between 10 and 11.

related to opw-706352
closes #26219