[FIX] account: unproxy company currency
This indirection is the only place in the whole codebase, which indirects in
this way.

This block downstream code which relies on:
 - access to the line's company
 - while not necesarily having access to the account's company

(being line's company different than account's company)

Disclaimer: This use case is not officially supported.
However, this indirection does not make sense under the currently
supported model, neither.