[IMP] hr_holidays: add kanban views for leave and allocation
Purpose of this commit is to add kanban views for allocation and leave
request. It is already possible to reach those views notably using the
activities systray. Default view contains only the name which is quite
sad. We therefore add real kanban views for both models.

This commit is linked to task ID 1876795 and PR #26571.