[FIX] account: reconciliation tour, set the url to use in python
Continuation and ending of 53978a5.

Before this commit, the bank statement reconciliation still sometimes
crashed.
Each step of the tour waits 10 secs before timeout and in this case,
changing URL may take more time since this commit 88c7348a because
we added 'debug=assets'. It was probably a mistake.

So, removing this from the URL would probably solve the error but
the second step would always be dependent on the first.
It's why we decided to load the tour with the correct URL defined
in Python.

Co-authored-by: Adrien Dieudonné <adr@odoo.com>