[FIX] *: fix UI glitch of kanban dropdown.
purpose of this commit is to show dropdown inside the kanban box with proper UI

By default, bootstrap use Popper.js for dynamic positioning. so, We need to
disable this issue with display static for removing glitch.

expected behaviour:
after applied this commit, kanban dropdown position will be inside of kanban box.

closes odoo/odoo#27150

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>