[FIX] base: correctly detect when admin logs in
Since 09e11e2b7039e4776504437af728a0f869e803e8 it's not possible to log in as user 1.

However, we reset the base url when the admin logs in, which is now user 2.

This fixes wkhtml2pdf was not able to load css file, and probably some other side effects.

PR: none
Task: 1879620