[IMP] mass_mailing : Add dependance with contacts
Contacts is now installed when installing mass_mailing.
It was possible to send mass mailing to contact without having
the contact app installed. which was a non sense because the user could not
manage the contacts and the admin could only send mail to the contact demo data,
which is not very useful.