[IMP] crm : add is partner blacklisted in crm.lead
Since the mass mailing blacklist, it's possible to see if a contact is blacklisted
using the blacklist_mixin.is_blacklisted. As crm.lead.opportunity view is displaying
one of the two fields email_from and partner_address_email and is_blacklisted is
computed on email_from, it was necessary to add a related fields that read
the is_blacklisted value of the related partner to see directly of the opportunity
is blacklisted or not.