[FIX] base: RNG validate text attribute for field (url)
The UrlWidget supports a text attribute to define the text of the <a> tag, but the text attribute wasn't supported in the RNG of the field.
Thus it wasn't possible to use it in tree views. However it already worked fine in form views because those are not RNG validated.

PR: none
Task: none