[FIX] mail: use contact app icon for partner-related activities in systray
As res.partner model is defined in base systray takes the base icon as icon
used in the systray when classifying activities by model. As base module has
no icon it leads to a blank icon being displayed in the systray. Purpose of
this commit is to use the icon from Contact application instead.

As contacts application now depends on mail we do it simply through
inheriting the systray_get_activities method that return the data to display
the systray elements.

This commit is linked to task ID 1882545 and PR #26797.