[FIX] stock_landed_costs: do not prorate
- Set the product A in FIFO, real-time valuation
- Receive 5 units @ 2000 each
- Create a landed cost of 1000, apply on the picking
  => the cost of A is 2200
- Sell one unit
- Create another landed cost of 1000, apply on the same picking

The cost of A is 2360, while one expects 2400.

The value of the move after the second landed cost is computed as:
value = previous value + prorata of landed cost 2
      = 11000 + (1000 * 4/5)
      = 11800

This is not correct. The whole value of the second landed cost should be
taken into account for the `value` field. The prorata is only useful for
the field `remaining_value`.

opw-1878229