[FIX] mail: toggle show/hide search options in discuss
Before this commit, the search option buttons were not
hidden when clicking on the zoom-in/zoom-out button next
to the searchview.

This is due to updating the control panel with the list
of buttons as the searchview buttons, instead of the
container. Since the visibility of the searchview buttons
is defined on the container, the toggler button and the
search view options lost synchronization.

This commit fixes the issue by providing the container
that is shared between search option buttons.

Task-ID 1877502