[FIX] web: use read to update record values after resequence
Commit:
https://github.com/odoo/odoo/commit/04c6089467590a8146ffd7d07024d96ada036e21
Updated the resequence logic in a way that was almost falser than before.
The added test however did work by coincidence, as index values were equal to
the sequence field values.
To be sure that we synchronize with what happens on the server, we do a read
after the resequence.
Additionnally, we take into account the result of the server resequence return;
if it is false, it means no resequencing happened, so we should not do a read.

opw 1867049

closes odoo/odoo#27184