[FIX] mail: filter messages in thread (Discuss)
Revision on https://github.com/odoo/odoo/commit/0703e0928750455a3e9b88d0bf0e140876779596

The commit above changes the way threads are rendered: the template
now uses a thread object, instead of a list of messages.

However, it was not handling messages that have been filtered from
the searchview. As a result, it was not possible to filter messages
at all.

This commit re-adds the possibility to filter messages, by providing
information on the domain of the thread in the template.

Task-ID 1877502