[REF] crm: don't archive but mark as lost
People don't understand archive. According to POs leads are lost and restored
and not archived / unarchived. Let us restore the mark as lost behavior.

As sometimes I am better than a goldfish let us point to this old task from
3 yeas ago: https://github.com/odoo/odoo/commit/1bb6eeeebe4358e6b3fcea120c2143d9332fccc9. This commit basically
reverts that one. It was linked to task ID 20996 that explainedwhy archive /
unarchive had to be used for leads as people don't understand won / lot for
leads.

This commit is linked to task ID 1880051 and PR #26977.