[FIX] l10n_eu_service: remove duplicate service tax rate
Since e590afc an error is logged when trying to insert duplicates in
ir_model_data. An error appears when installing this module because of
the duplicate.

As this module is not yet tested, this error flew under the radar.

Note: a new test that installs l10n* is developped