[FIX] web: tests: adapt JS tests from forward-port 9fe70007019fd6dcabb3652ff5abdf9bd819ee2a
The test was failing because it was using an old approach to mock
services in JS tests. Also, it provides ajax service in the list
of services, which was not supported in a mocked environment.

With this commit, when an ajax service is provided in the test,
it will use this service instead of using mocked RPCs from the
mocked server.