[CHG] *: "family" names of admin & demo users
avw/fp? request

9f7bb04324d7b713ad17d93474cca5f05e11fdf4 changed various demo data,
including putting more "realistic" names in for the administrator and
demo users in demo mode. This is apparently inconvenient when actually
demo-ing Odoo as it's harder to see at a glance whether the current user
is the administrator or "regular" (demo) user of the db, though this
should have less impact since the userification of the admin
(09e11e2b7039e4776504437af728a0f869e803e8).