[FIX] hr_timesheet: bind format/parse method for timesheet_uom widget
The `timesheet_uom` is a fake widget: it is bind in field widget
registry on other existing widget depending on the session_info
preferences (e.i.: float_time, float_toggle, ...). Therefore, it
does not have formatter/parser methods in the field utils method.

This commit binds the formatter and the parser of the widget on which
`timesheet_uom` is bind, to make it coherent. This avoid having a
traceback on pivot view when the `timesheet_uom` widget is set on
a field node.