[FIX] ir_ui_view: do not check arch twice upon install/upgrade
After a module upgrade, validate the architecture of the module's views that
are impacted by updates, but have not been checked yet.  Before this patch,
views were checked twice on average.

This patch speeds up the installation of modules by about 15% without demo
data.  With demo data, the speedup is around 10%.