[FIX] stock: package to master data
commit ac5a142bb98b5ed730d5faddd428568a0c132f2d moved package menu to
operation and rename it delivery package. However move it to opeation
could confuse the user and let him imagine that he is able to create the
package content from this menu which is not the case. This menu only
allow you to see the package, its location and its content.

Also package are not always delivery package since it could be use in
internal picking type so it's better to remove the 'delivery' from the
menu item.