[IMP] hr_holidays: department leaves should be based on members
Leaves and allocations have a My Department filter allowing to see leaves
or allocations linked to my department. Currently it is defined as the
department the current user manages. It is better to see the departments
current user is a member as it has more uses notably for HR officers.

Coming from @est-odoo direct live feedback.