[FIX] web: 'Shortcuts' in user menu (community)
Revision on https://github.com/odoo/odoo/commit/661580259363290e4dab06d3cfa641430012f02a

This menu item should not appear in the community version of Odoo.
The commit above mistakenly added this menu item, even though no
logic was implemented for it. This was a consequence of the copy
& paste approach to rapidly deploy the new navbar in community.

(Note: a big refactoring of the web client is coming very soon,
so this duplicated code issue should only be temporarily).

This commit fixes the issue by removing this menu item in the
community version.