[IMP] hr_{attendance,recruitment,holidays,expense,contract}: add activity view on main actions
Purpose is to extend the use of the newly-introduced activity view allowing
to see at a glance activities to perform on a given model. It eases daily
job of people working with activities.

Updated

  * employee main menu actions;
  * leave and allocation request main menu actions;
  * applicant main menu actions;
  * expense and expense report main menu actions;
  * contract main menu action;
  * attendance employee main action;

Following @est-odoo advice the Next Activities menu entry in hr_recruitment
is removed as there are already a lot of ways of managing activities: using
the systray, using activity view present on most actions, ...

This commit is linked to task ID 1889413.