[FIX] account, portal : align invoices counter and view domains
The idea of removing the domain was to have the portal views based
only on access rights and rules. We were just not really consistent
in not removing the domain on the counter. If the user has the right
to see also supplier invoices, it makes sens to let the user see
them in the portal. The domain on the counter is then here removed
to align both counter and views.

Fix commit eca38c4dc50387e180e0b46e71dcecd696564186
Closes PR #27583
Link to Task ID 30985