[FIX] auth_password_policy: Cannot read property 'decorations' of undefined
Revision on https://github.com/odoo/odoo/commit/0c5e617ff7ecfa0114a891b0dcd5b23d8389b801

The commit above introduced a visual 'password strength' next to
the password wizard in Settings > Users, when the module
`auth_password_policy` is installed.

However, there was an traceback when trying to access the 'General
Settings' when some specific apps and modules were installed.

This issue comes from the small design issue with the new
PasswordField: it was designed as a field object that was
instantiated and returned in the `init` function of an InputField
with the password attribute set. This is not a solid design,
because the documentation of the `init` function does not
specify that it should return the newly instantiated object.

As a result, the implementation was very fragile and only worked
on FieldChar but not FieldText.

This commit fixes the issue by turning the FieldPassword a field
widget. As a result, the enabling of the password meter next to
a password input has changed:

Before this commit:

```
<field name="new_passwd" password="True" options="{'password_meter': True}"/>
```

With this commit:

```
<field name="new_passwd" password="True" widget="password_meter"/>
```

Closes #27440

closes odoo/odoo#27586